Fix bug where the configuration widget code would always run due to exporting elements from it's file
